// src/app/shared/enums/auth-error-messages.enum.ts
export enum AuthErrorMessages {
    // firebase errors
    USER_NOT_FOUND = 'USER_NOT_FOUND',
    USER_IS_NOT_SIGNED_IN = 'USER_IS_NOT_SIGNED_IN',
    WRONG_PASSWORD = 'WRONG_PASSWORD',
    USER_IS_DISABLED = 'USER_IS_DISABLED',
    INVALID_EMAIL = 'INVALID_EMAIL',
    WEAK_PASSWORD = 'WEAK_PASSWORD',
    EMAIL_ALREADY_IN_USE = 'EMAIL_ALREADY_IN_USE',
    NETWORK_REQUEST_FAILED = 'NETWORK_REQUEST_FAILED',
    INTERNAL_ERROR = 'INTERNAL_ERROR',
    // non-firebase errors
    UNVERIFIED_EMAIL = 'UNVERIFIED_EMAIL',
    UPDATING_EMAIL_FAILED = 'UPDATING_EMAIL_FAILED',
    SENDING_PASSWORD_RESET_EMAIL_FAILED = 'SENDING_PASSWORD_RESET_EMAIL_FAILED',
    REGISTRATION_FAILED = 'REGISTRATION_FAILED',
    ACCOUNT_CREATION_IS_DISABLED = 'ACCOUNT_CREATION_IS_DISABLED',
    PASSWORD_RESETTING_IS_DISABLED = 'PASSWORD_RESETTING_IS_DISABLED',
    VERIFICATION_EMAIL_REQUESTING_IS_DISABLED = 'VERIFICATION_EMAIL_REQUESTING_IS_DISABLED',
}